+++ src/Commands/Fun/Trivia/Trivia.ts
@@ -1,9 +1,10 @@
 import { Command } from "../../../Structures/Command";
-import { Message } from "discord.js";
-import { trivia } from "../../../lib/Backend/Trivia/Trivia";
+import { Message, GuildMember } from "discord.js";
+import { Trivia, categoryRegex, categories } from "../../../lib/Backend/Trivia/Trivia";
 import { shuffle } from '../../../lib/Utility/Array';
 
-const unbase64 = (base64: string) => Buffer.from(base64, 'base64').toString();
+type diff = 'easy' | 'medium' | 'hard';
+const games: { [key: string]: string } = {};
 
 export default class extends Command {
     constructor() {
@@ -23,68 +24,94 @@ export default class extends Command {
     }
 
     async init(message: Message, args: string[]) {
-        const list = await trivia.fetchList();
+        if(message.guild.id in games) {
+            return message.channel.send(this.Embed.fail(`A game is already going on in this guild!`));
+        }
+
+        const list = await Trivia.fetchList();
         if(!list) {
             return message.channel.send(this.Embed.fail('An unexpected error occurred!'));
         }
 
-        const category = args.join(' ').match(trivia.categoryRegex);
-        if(category.length === 0) {
-            return message.channel.send(this.Embed.fail('Invalid category provided! Use the ``triviacategory`` command!'));
+        const category = Number.isInteger(+args[0])
+            ? categories.filter(d => d.id === +args[0]).shift()?.name
+            : args.join(' ').match(categoryRegex)?.shift();
+                    
+        const cid = category ? categories.filter(e => e.name.toLowerCase() === category.toLowerCase()) : [];
+        if(category?.length === 0 || cid.length === 0) {
+            return message.channel.send(this.Embed.generic('Invalid category provided! Use the ``triviacategory`` command!'));
         }
 
-        const [difficulty, q] = args.slice(category[0].split(' ').length);
-        if(!difficulty || ['easy', 'medium', 'hard'].indexOf(difficulty.toLowerCase()) === -1) {
-            return message.channel.send(this.Embed.fail('Invalid difficulty provided!'));
+        const [difficulty, q] = args.slice(Number.isInteger(+args[0]) ? 1 : category.split(' ').length);
+        if(!difficulty || !['easy', 'medium', 'hard'].includes(difficulty.toLowerCase())) {
+            return message.channel.send(this.Embed.generic('Invalid difficulty provided!'));
         }
 
-        if(isNaN(+q)) {
-            return message.channel.send(this.Embed.fail('Invalid amount of questions!'));
+        if(isNaN(+q) || !Number.isInteger(+q)) {
+            return message.channel.send(this.Embed.generic('Invalid amount of questions!'));
         }
 
-        const questions = await trivia.fetchQuestions(+q > 10 ? 10 : +q, category[0], difficulty);
-        shuffle(questions.results); // modifies original array
+        const questions = await Trivia.fetchQuestions(+q > 10 ? 10 : +q, cid.shift().id, difficulty as diff);
         
-        let sent: Message;
-        const winner: { [key: string]: number } = {};
+        const guesses: { [key: number]: string[] } = {}
+        const winner: { [key: number]: GuildMember } = {};
+        let msg: Message = null;
+        games[message.guild.id] = message.channel.id;
 
-        for(const question of questions.results) {
-            if(!super.hasPermissions(message)) {
+        for(const question of questions) {
+            if(msg?.deleted) {
+                delete games[message.guild.id];
                 return;
             }
-
-            const answers = shuffle(question.incorrect_answers.concat(question.correct_answer)); // answers.indexOf(correct answer)
-            const embed = this.Embed.success()
-                .setTitle(`Question ${questions.results.indexOf(question) + 1}`)
-                .setDescription(`
-                \`\`${unbase64(question.question)}\`\`
-                
-                ${answers.map((a, i) => `\`\`${i + 1}\`\`: ${unbase64(a)}`).join('\n')}
-                `);
             
-            if(!sent || sent.deleted || !sent.editable) {
-                sent = await message.channel.send(embed);
-            } else if(sent instanceof Message) {
-                await sent.edit(embed);
+            const answers = shuffle([question.correct_answer, ...question.incorrect_answers]);
+            const index = questions.indexOf(question);
+            const embed = this.Embed.success()
+                .setTitle(`${question.category} - ${question.difficulty}`)
+                .setDescription(`
+                \`\`${question.question}\`\`
+                Answers:
+                ${answers.map((a, i) => `**${i + 1}:** ${a}`).join('\n')}
+                `);
+
+            msg = msg ? await msg.edit(embed) : await message.channel.send(embed);
+
+            const filter = (m: Message) =>  {
+                if(guesses[index]?.includes(m.author.id)) {
+                    return false;
+                }
+
+                guesses[index] ? guesses[index].push(m.author.id) : (guesses[index] = [m.author.id]);
+                setTimeout(() => guesses[index].splice(guesses[index].indexOf(m.author.id), 1), 1000);
+
+                return m.content?.toLowerCase() === question.correct_answer.toLowerCase().trim()
+                       || +m.content === answers.indexOf(question.correct_answer) + 1
             }
 
-            const filter = (m: Message) => m.content?.toLowerCase() === unbase64(question.correct_answer).toLowerCase().trim()
-                                           || +m.content === answers.indexOf(question.correct_answer) + 1
-            const collected = await message.channel.awaitMessages(filter, { max: 1, time: 30000 });
+            // If two people type the correct answer at near the same time and two questions have the same correct answer 
+            // it instantly skips over the second question. - Pseudo
+            await new Promise(r => setTimeout(r, 1000));
+            const collector = message.channel.createMessageCollector(filter, {
+                time: 30000
+            });
 
-            if(collected.size === 1) {
-                const guesser = collected.first().member;
-                winner[guesser.id] = winner[guesser.id] ? winner[guesser.id] + 1 : 1;
-            }
+            await new Promise(r => {
+                collector.on('collect', (m: Message) => {
+                    winner[index] = m.member;
+                    collector.stop();
+                });
+
+                collector.on('end', r);
+            });
         }
 
-        const w = Object.entries(winner).sort((a, b) => b[1] - a[1]).shift();
-        const user = w[1] === 0 ? null : await message.client.users.fetch(w[0]);
+        delete games[message.guild.id];
+        const won = Object.values(winner).sort((a, b) =>
+            Object.values(winner).filter(v => v.id === a.id).length - Object.values(winner).filter(v => v.id === b.id).length
+        ).pop();
 
-        if(sent.deleted || !sent.editable) {
-            return message.channel.send(this.Embed.success(`${user ?? 'No one'} won with ${w[1]} correct answers.`));
+        if(!msg.deleted) {
+            return msg.edit(this.Embed.success(`${won ?? 'No one'} won the game!`));
         }
-
-        return sent.edit(this.Embed.success(`${user ?? 'No one'} won with ${w[1]} correct answers.`));
     }
 }