+++ CHANGELOG.md
@@ -234,3 +234,8 @@
 * Remove ``copy`` script because it's no longer needed.
 * Fixed ``build`` script not removing the ``build`` directory on linux.
 * Import JSON using native TypeScript abilities now that it works.
+* Trivia command has been re-written.
+    * Fixes a bug where if two people answer a question in close proximity of time and the next question has the same answer, it won't be instantly answered.
+    * Fixed a bug where a person could input a non-integer number of questions.
+    * Questions are cached, and 50 of them will always be fetched from the API. Should help performance over multiple games.
+    * Fixed a bug where you couldn't use the number of a category to start a game.